refactor(racingcar): change package structure and match test to main
