+++ src/main/java/racingcar/domain/GameManager.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 import racingcar.domain.Car;
 import racingcar.view.InputView;
