+++ src/test/java/racingcar/domain/GameManagerTest.java
@@ -1,9 +1,10 @@
-package racingcar;
+package racingcar.domain;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 import racingcar.domain.Car;
+import racingcar.domain.GameManager;
 
 import java.lang.reflect.Method;
