+++ src/test/java/racingcar/util/RandomUtilsTest.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.util;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
