+++ src/test/java/racingcar/util/ValidatorUtilsTest.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.util;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
